Refactor to avoid circular imports when using serialize.py externally.
View classes no longer accessible from `mentions.views` - now imported
from their individual modules explicitly.

serialize.py moved into views package.
