--- mentions/views/__init__.py
@@ -1,3 +1,0 @@
-from .dashboard import WebmentionDashboardView
-from .retrieve_mentions import GetMentionsByTypeView, GetMentionsView
-from .submit_mentions import WebmentionView
