+++ mentions/views/submit.py
@@ -1,4 +1,5 @@
 import logging
+from typing import Optional
 
 from django.http import HttpResponseBadRequest
 from django.shortcuts import render
@@ -25,7 +26,13 @@ class WebmentionView(View):
 
     def get(self, request):
         form = SubmitWebmentionForm()
-        return render(request, "mentions/webmention-submit-manual.html", {"form": form})
+        return render(
+            request,
+            "mentions/webmention-submit-manual.html",
+            {
+                "form": form,
+            },
+        )
 
     def post(self, request):
         log.debug("Receiving webmention...")
@@ -48,7 +55,7 @@ class WebmentionView(View):
         return render(request, "mentions/webmention-accepted.html", status=202)
 
 
-def _get_client_ip(request):
+def _get_client_ip(request) -> Optional[str]:
     x_forwarded_for = request.META.get("HTTP_X_FORWARDED_FOR")
     if x_forwarded_for:
         ip = x_forwarded_for.split(",")[-1].strip()
